Fix attach_perf_event error.
This is caused data length of argument between python <-> libbpf different on
ARM32. The arguments sample_period / sample_freq of perf_event_open() system
call are u64, the period in python 2.7 for bcc is int, so the length is 32
bits wide, and this caused the issue, need to fix it in _init_.py.