biolatency.py Add -j flag for json output
In it's current form, storing the histogram data and analyzing the same for a large set of machines is difficult.
With this flag, it is possible to store the histogram as a dictionary/json file which can be imported easily to be analyzed with libraries like Pandas. An example output looks like below:

{
    'ts': '2020-12-30 14:49:37',
    'val_type': 'msecs',
    'data':
        [
            {'interval-start': 0, 'interval-end': 1, 'count': 2},
            {'interval-start': 2, 'interval-end': 3, 'count': 1}
        ],
    'flags': 'Sync-Write'
}

Change-Id: I439e9bdeda6a7f8204ece9ff7fe114a98f7d0639