libbpf-tools: offcputime: Ensure volatile variable size meets 64-bit requirement
The skeleton code generated by bpftool  uses 'static_assert' to verify
that the size of a volatile variable conforms to that of a 64-bit
system. To satisfy this check, this patch use 'long long' which is 8
bytes on both 32-bit and 64-bit systems.

In eBPF subsystem, the kernel function 'stack_map_alloc' would enforce
8-byte alignment of map size. This patch ensures that
'perf_max_stack_depth' is rounded up to meet the 8-byte alignment
requirement if it is not already aligned.

Change-Id: Ie5edc4feb13a28f584ae6adb2d44d603f9f3d149
Signed-off-by: roger_weijie <roger_weijie@realtek.com>