libbpf-tools/futexctn: add futex_time32 support
This change adds support for the `futex_time32` syscall to the `futexctn`
tool. This is necessary because many 32-bit applications use the `futex`
syscall with a 32-bit time structure.

To avoid code duplication, the existing BPF code has been refactored to
use `__always_inline` functions for both `futex` and `futex_time32`
syscall tracing. This makes the code cleaner and easier to maintain.

Signed-off-by: Ism Hong <ism.hong@realtek.com>