libbpf-tools/syscount: Add syscall table for arm32
Add a syscall table to translate syscall numbers to names for the arm32
architecture. This is necessary for syscount to properly display
syscall information on arm32 systems.

A new `-A` (`--arm32`) option is added to `syscount` to allow users
to explicitly request the use of the arm32 syscall table. This is
useful for tracing 32-bit ARM applications running on an aarch64
system, as the syscall numbers for 32-bit and 64-bit ARM
architectures are different.

Signed-off-by: Ism Hong <ism.hong@realtek.com>