libbpf-tools/execsnoop: Fix incorrect argv access on ARM32
On ARM32, accessing argv with `&argv[i]` can lead to an incorrect
address calculation. This is because the compiler might not generate the
correct code to handle the pointer arithmetic for `const char *const
*argv`.

This patch fixes the issue by manually calculating the address of each
argument in `argv` using integer arithmetic. This ensures that the
correct address is read, and the arguments are correctly fetched.

Change-Id: Id7f7c84594dbdbb56d21638437c76626526290d2
Signed-off-by: Ism Hong <ism.hong@realtek.com>