libbpf-tools/syncsnoop: Use kprobe for sync syscall
The 'ksyscall' BPF program type incorrectly attaches to '__se_sys_sync'
on arm32 platforms for the 'sync' syscall. This fails because 'sync'
is a zero-argument syscall, and its correct function name is 'sys_sync'.

To resolve this, attach to the 'sync' syscall using a 'kprobe' on
'sys_sync' directly. This ensures the BPF program is attached to the
correct kernel function.

Change-Id: I0a2c8e7bbad358d5025057245ad3bab7b65ec4c6