libbpf-tools: Set stdout to line-buffered
The printf() function buffers output in blocks when stdout is
redirected to a file or pipe. This is not ideal for tracing tools,
where real-time event visibility is important.

This commit adds a constructor function to trace_helpers.h that
calls setvbuf(stdout, NULL, _IOLBF, 0). This forces stdout to be
line-buffered for all tools that include this header, ensuring that
trace output is flushed after each newline.

Change-Id: I0aecb4fca4de0f158e6fca92f93678e64f75b073
Signed-off-by: Ism Hong <ism.hong@realtek.com>