libbpf-tools/smctop: Use __data_loc for irq name in smctop
The smctop tool was failing with an ftrace message error when tracing
the irq/irq_handler_entry event. This was because accessing the irq name
field for this tracepoint requires using the `__data_loc` macro to
properly resolve the data location.

This commit fixes the issue by using `__data_loc` to access the irq
name, ensuring that the smctop tool can correctly trace IRQ handler
events.

Change-Id: Ic69695e1d17c0947e4f3aa569ffd382966fc980f
Signed-off-by: ism.hong <ism.hong@realtek.com>