libbpf-tools/forksnoop: Fix incorrect PPID/PID in fork events
The forksnoop tool was reporting incorrect PPID and PID values for fork
events, causing discrepancies with the output of the original bcc Python
version.

This was due to incorrect data retrieval in the
`tracepoint__sched_process_fork` BPF program. The PPID was being read
from the great-grandparent process and the PID from the grandparent
process, instead of the grandparent and parent, respectively.

This commit corrects the logic to match the bcc implementation, ensuring
consistent and accurate reporting of process hierarchy information.  The
PID filtering logic is also corrected to filter only on the parent's
PID, as in the original tool.

Change-Id: I3e0106d38e063dfc0301f7dd3ed73068bfd3098b
Signed-off-by: Ism Hong <ism.hong@realtek.com>