libbpf-tools/cpuidle: Fix map key type and align state names
The data type of the key for the 'all_cpu_sleep' BPF map was incorrectly
specified as '__u32' in the user-space lookup, while the map itself is
defined with a '__u64' key. This commit corrects the key's data type to
'__u64' to prevent potential issues.

Additionally, the idle state naming logic has been updated to always
append a numerical suffix (e.g., "-0", "-1"). This ensures that the
output is consistent with the Python version of 'cpuidle' and improves
output alignment.

Change-Id: Ic6dd292a242ca32b87a6883dc7b43e1a5d4da1cf
Signed-off-by: ism.hong <ism.hong@realtek.com>