libbpf-tools/cputop: fix wrong PREEMPT_CNT issue
The logic for detecting task preemption in `cputop` was based on an
incorrect task state mask. The previous implementation used a mask of
0x100, which does not accurately represent a preempted task state.

This commit corrects the preemption detection by using `TASK_STATE_MAX`,
which is the correct mask for identifying preempted tasks. This ensures
that the `PREEMPT_CNT` metric reported by `cputop` is accurate.

Change-Id: I7964b7f36b8edbab17c86e3182a937e834004815
Signed-off-by: Ism Hong <ism.hong@realtek.com>