libbpf-tools/rtkheaptop: align CALLER field to BCC version
The `CALLER` field in the libbpf version of `rtkheaptop` was not
displaying the `tgid-*` information when the kernel did not provide a
caller name. This was inconsistent with the BCC Python version of the
tool, which provides the `tgid` as a fallback.

This commit aligns the behavior of the libbpf tool with the BCC
version by implementing a helper function, `get_caller_str`, which
checks for an empty caller name and substitutes it with the `tgid` in
the format "tgid-<tgid>".

This change ensures that the output of `rtkheaptop` is consistent
across both versions, providing more informative output to the user.

Change-Id: I6ab34cd59627add3251a838e11902dd0baa46613
Signed-off-by: Ism Hong <ism.hong@realtek.com>